src/utils.py
"""
Utility functions for the Job CV Customization System.
Includes LinkedIn scraping, CV parsing, and PDF generation.
"""

import yaml
from pathlib import Path
from typing import Dict, Optional
from datetime import datetime


class CVParser:
    """Utilities for parsing and working with CV files"""
    
    @staticmethod
    def load_cv(cv_path: str) -> dict:
        """Load CV from YAML file"""
        with open(cv_path, 'r') as file:
            return yaml.safe_load(file)
    
    @staticmethod
    def save_cv(cv_data: dict, output_path: str) -> None:
        """Save CV to YAML file"""
        Path(output_path).parent.mkdir(parents=True, exist_ok=True)
        with open(output_path, 'w') as file:
            yaml.dump(cv_data, file, default_flow_style=False, sort_keys=False)
    
    @staticmethod
    def cv_to_string(cv_data: dict) -> str:
        """Convert CV dict to YAML string"""
        return yaml.dump(cv_data, default_flow_style=False, sort_keys=False)
    
    @staticmethod
    def string_to_cv(cv_string: str) -> dict:
        """Convert YAML string to CV dict"""
        return yaml.safe_load(cv_string)


class LinkedInScraper:
    """
    Utilities for scraping LinkedIn job postings.
    NOTE: This is a placeholder - actual LinkedIn scraping requires handling:
    - LinkedIn's terms of service
    - Authentication
    - Dynamic content loading
    - Rate limiting
    """
    
    @staticmethod
    def scrape_job_posting(job_url: str) -> Optional[str]:
        """
        Scrape a LinkedIn job posting
        
        Args:
            job_url: URL to the LinkedIn job posting
            
        Returns:
            Job posting text content
            
        Note:
            This is a placeholder. Real implementation would need:
            - Selenium or Playwright for dynamic content
            - LinkedIn login credentials
            - Proper error handling and rate limiting
        """
        print(f"[TODO] Implement LinkedIn scraping for: {job_url}")
        print("This requires handling LinkedIn's authentication and dynamic content")
        return None
    
    @staticmethod
    def extract_job_text_from_html(html_content: str) -> Optional[str]:
        """
        Extract job posting text from HTML
        
        Args:
            html_content: Raw HTML from LinkedIn job page
            
        Returns:
            Extracted job posting text
        """
        # Placeholder for BeautifulSoup implementation
        print("[TODO] Implement HTML parsing for job postings")
        return None


class PDFGenerator:
    """Generate PDF from customized CV"""
    
    @staticmethod
    def cv_yaml_to_pdf(cv_data: dict, output_path: str = None) -> str:
        """
        Convert CV from YAML/dict format to PDF
        
        Args:
            cv_data: CV dictionary
            output_path: Where to save the PDF
            
        Returns:
            Path to generated PDF
            
        Note:
            Placeholder for PDF generation using reportlab.
            Real implementation would format the CV nicely for PDF output.
        """
        if output_path is None:
            output_dir = Path("outputs")
            output_dir.mkdir(exist_ok=True)
            timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
            output_path = output_dir / f"cv_{timestamp}.pdf"
        
        print(f"[TODO] Generate PDF: {output_path}")
        print("Implementation would use reportlab to format CV as professional PDF")
        
        return str(output_path)
    
    @staticmethod
    def generate_professional_pdf(cv_data: dict, output_path: str) -> str:
        """
        Generate a professionally formatted PDF from CV data
        
        Future implementation will include:
        - Professional formatting with proper spacing
        - Color scheme and typography
        - Logo or header with contact info
        - Proper page breaks
        - ATS-friendly formatting
        """
        print(f"Generating professional PDF: {output_path}")
        # Will use reportlab library
        return output_path


class JobPostingMocker:
    """Mock job postings for testing without LinkedIn"""
    
    @staticmethod
    def get_sample_postings() -> Dict[str, str]:
        """Get sample job postings for testing"""
        return {
            "data_scientist": """
                Senior Data Scientist
                Company: DataCorp Inc.
                Location: New York, NY
                
                We're seeking a Senior Data Scientist to lead our analytics team.
                
                Responsibilities:
                - Build predictive models and analytical frameworks
                - Lead cross-functional projects with engineering and product teams
                - Mentor junior data scientists
                - Deploy models to production using Python and cloud platforms
                
                Requirements:
                - 5+ years in data science or analytics
                - Expert in Python, SQL, and machine learning libraries
                - Experience with TensorFlow or PyTorch
                - Knowledge of AWS or GCP
                - Strong communication skills
                
                Nice to Have:
                - Published research or papers
                - Experience with A/B testing and experimentation
                - Knowledge of big data tools (Spark, Hadoop)
            """,
            
            "backend_engineer": """
                Senior Backend Engineer
                Company: ScaleTech
                
                Join our backend team building scalable systems.
                
                Responsibilities:
                - Design and implement backend services
                - Improve system performance and reliability
                - Lead architectural decisions
                - Mentor engineers on best practices
                
                Requirements:
                - 5+ years backend development experience
                - Proficiency in Python or Java
                - Experience with microservices architecture
                - Strong knowledge of databases (SQL and NoSQL)
                - AWS or similar cloud platform experience
                
                Nice to Have:
                - Kubernetes experience
                - Experience with message queues (RabbitMQ, Kafka)
                - Open source contributions
            """,
            
            "full_stack_engineer": """
                Full Stack Engineer
                Company: WebInnovate
                
                Build modern web applications end-to-end.
                
                Responsibilities:
                - Develop full-stack features using React and Node.js
                - Collaborate with designers and product managers
                - Write clean, maintainable code
                - Participate in code reviews
                
                Requirements:
                - 3+ years full stack development
                - Strong JavaScript/TypeScript skills
                - React or Vue.js experience
                - Node.js or similar backend framework
                - PostgreSQL or MongoDB experience
                
                Nice to Have:
                - Docker and CI/CD experience
                - AWS or Firebase experience
                - Mobile development experience
            """
        }


# Example usage
if __name__ == "__main__":
    # Test CV parsing
    cv = CVParser.load_cv("base_cv.yaml")
    print("✓ CV loaded successfully")
    print(f"  Name: {cv['personal_info']['name']}")
    
    # Test job posting mocking
    postings = JobPostingMocker.get_sample_postings()
    print(f"\n✓ Available sample postings: {list(postings.keys())}")
