requirements.txt
crewai==0.30.0
crewai-tools==0.1.6
python-dotenv==1.0.0
pydantic>=2.6.1
PyYAML==6.0.1
anthropic==0.25.0
requests==2.31.0
beautifulsoup4>=4.12.3
selenium>=4.18.1
reportlab==4.0.7
pydantic-settings==2.1.0
